Finished size() and prepend() function
size() iterates through the linked list starting
from the head of the list and returns a count of
each node object in the linked list. prepend()
makes two temporary varibles, one representing
the head, and one representing our new created node
with the parameter value in it. We set the next node
of our new created node to point to the previous
head node, then we assign the new node to be the
NEW head node since we are inserting at the
start of the list.