Finished getHead() and getTail() functions
Had to change their head() and tail() original
function names to getHead() and getTail() to
avoid naming conflicts with the linkedlist
variable "head" and "tail". getHead() just returns
the head variable since it already contains
the start of the list node, and tail does iterate
through the list to get to the end, and returns
the last node in the linked list.