finished the at(index) function
Made the at() function with all of the
edge cases and pitfalls like out of bounds and
if the linked list is empty. Since the function
gets the index from the parameter, we use a
for loop after confirming that the above conditions
were fine (we can assume that the given index is
within bounds if it passes the previous pitfall
checks) and loop through the linked list
starting at the head until we get to
the index, then return the node.