Finished find(value) function
find() function uses the exact same logic as the
contains() function, but with a index counter
that increments as we search through the linked
list for the node with the input value in it.
returns index of the found node, or null if
the node with the input value is not found
or if the list is empty.