Finished extra credit functions
insertAt() and removeAt() use similar logic in that
they both check multiple edge cases like
if the index was out of bounds of the list, or
if the index given was the end index or
the start index. For inserting, it is adding
a node to the linked list at that index so
the node before it needs to link to the
new node, and the new node needs to link to
the node that was there before the insert.
For removing, the only change that was needed
was to make sure that the previous node before
the removed node points at the next node
that came after the node being removed.

Essentially, these operations required temporary
variables that pointed to the previous
node to re-wire their nextNode property to the
proper node after insertion/deletion. Tested
it all in index.js and main.js, and they are working
as intended.